fix: unblock PR #4214 CI — split widgetbook catalog and registry parity
Extract naval mission Widgetbook stories into catalog_dialogs_naval_mission.dart
to stay under the 1000 non-comment line cap, and register DLG31001–DLG31003 in
the UiScreenIds registry parity test.

Refs #4213
