+++ app/test/ui_screen_ids_registry_parity_test.dart
@@ -54,6 +54,9 @@ const Map<String, String> _expectedConstants = {
   'newGameLeaderSelectionDialog': UiScreenIds.newGameLeaderSelectionDialog,
   'moveArmyDialog': UiScreenIds.moveArmyDialog,
   'moveFleetDialog': UiScreenIds.moveFleetDialog,
+  'navalMissionMenuDialog': UiScreenIds.navalMissionMenuDialog,
+  'navalMissionTargetDialog': UiScreenIds.navalMissionTargetDialog,
+  'navalMissionFleetPickerDialog': UiScreenIds.navalMissionFleetPickerDialog,
   'transferToHomeFleetDialog': UiScreenIds.transferToHomeFleetDialog,
   'turnNewsDialog': UiScreenIds.turnNewsDialog,
   'nextTurnConfirmation': UiScreenIds.nextTurnConfirmation,
